Add files via upload
Added  try/catch blocks to LoadCars method and try/catch blocks to Main method in order to bubble up the exceptions. Also assigned the Car Vin number using the Guid class when the Car class is instantiated, instead of within the  "foreach" loop in the Main method. The Vin number assignment only needs to be done once.